backend/requirements.txt
Flask==2.3.2
requests==2.31.0
gunicorn==21.2.0
boto3==1.26.162
python-dotenv